+++ src/Symfony/Component/Form/ChoiceList/Factory/CachingFactoryDecorator.php
@@ -19,6 +19,9 @@ use Symfony\Contracts\Service\ResetInterface;
 /**
  * Caches the choice lists created by the decorated factory.
  *
+ * To cache a list based on its options, arguments must be decorated
+ * by a {@see Cache\AbstractStaticOption} implementation.
+ *
  * @author Bernhard Schussek <bschussek@gmail.com>
  * @author Jules Pietri <jules@heahprod.com>
  */
@@ -80,25 +83,42 @@ class CachingFactoryDecorator implements ChoiceListFactoryInterface, ResetInterf
 
     /**
      * {@inheritdoc}
+     *
+     * @param callable|Cache\ChoiceValue|null  $value  The callable or static option for
+     *                                                 generating the choice values
+     * @param callable|Cache\ChoiceFilter|null $filter The callable or static option for
+     *                                                 filtering the choices
      */
-    public function createListFromChoices(iterable $choices, $value = null)
+    public function createListFromChoices(iterable $choices, $value = null/*, $filter = null*/)
     {
+        $filter = \func_num_args() > 2 ? func_get_arg(2) : null;
+
         if ($choices instanceof \Traversable) {
             $choices = iterator_to_array($choices);
         }
 
-        // Only cache per value when needed. The value is not validated on purpose.
+        $cache = true;
+        // Only cache per value and filter when needed. The value is not validated on purpose.
         // The decorated factory may decide which values to accept and which not.
         if ($value instanceof Cache\ChoiceValue) {
             $value = $value->getOption();
         } elseif ($value) {
-            return $this->decoratedFactory->createListFromChoices($choices, $value);
+            $cache = false;
+        }
+        if ($filter instanceof Cache\ChoiceFilter) {
+            $filter = $filter->getOption();
+        } elseif ($filter) {
+            $cache = false;
         }
 
-        $hash = self::generateHash([$choices, $value], 'fromChoices');
+        if (!$cache) {
+            return $this->decoratedFactory->createListFromChoices($choices, $value, $filter);
+        }
+
+        $hash = self::generateHash([$choices, $value, $filter], 'fromChoices');
 
         if (!isset($this->lists[$hash])) {
-            $this->lists[$hash] = $this->decoratedFactory->createListFromChoices($choices, $value);
+            $this->lists[$hash] = $this->decoratedFactory->createListFromChoices($choices, $value, $filter);
         }
 
         return $this->lists[$hash];
@@ -106,9 +126,18 @@ class CachingFactoryDecorator implements ChoiceListFactoryInterface, ResetInterf
 
     /**
      * {@inheritdoc}
+     *
+     * @param ChoiceLoaderInterface|Cache\ChoiceLoader $loader The loader or static loader to load
+     *                                                         the choices lazily
+     * @param callable|Cache\ChoiceValue|null          $value  The callable or static option for
+     *                                                         generating the choice values
+     * @param callable|Cache\ChoiceFilter|null         $filter The callable or static option for
+     *                                                         filtering the choices
      */
-    public function createListFromLoader(ChoiceLoaderInterface $loader, $value = null)
+    public function createListFromLoader(ChoiceLoaderInterface $loader, $value = null/*, $filter = null*/)
     {
+        $filter = \func_num_args() > 2 ? func_get_arg(2) : null;
+
         $cache = true;
 
         if ($loader instanceof Cache\ChoiceLoader) {
@@ -123,14 +152,20 @@ class CachingFactoryDecorator implements ChoiceListFactoryInterface, ResetInterf
             $cache = false;
         }
 
-        if (!$cache) {
-            return $this->decoratedFactory->createListFromLoader($loader, $value);
+        if ($filter instanceof Cache\ChoiceFilter) {
+            $filter = $filter->getOption();
+        } elseif ($filter) {
+            $cache = false;
         }
 
-        $hash = self::generateHash([$loader, $value], 'fromLoader');
+        if (!$cache) {
+            return $this->decoratedFactory->createListFromLoader($loader, $value, $filter);
+        }
+
+        $hash = self::generateHash([$loader, $value, $filter], 'fromLoader');
 
         if (!isset($this->lists[$hash])) {
-            $this->lists[$hash] = $this->decoratedFactory->createListFromLoader($loader, $value);
+            $this->lists[$hash] = $this->decoratedFactory->createListFromLoader($loader, $value, $filter);
         }
 
         return $this->lists[$hash];
@@ -138,6 +173,12 @@ class CachingFactoryDecorator implements ChoiceListFactoryInterface, ResetInterf
 
     /**
      * {@inheritdoc}
+     *
+     * @param array|callable|Cache\PreferredChoice|null $preferredChoices The preferred choices
+     * @param callable|false|Cache\ChoiceLabel|null     $label            The option or static option generating the choice labels
+     * @param callable|Cache\ChoiceFieldName|null       $index            The option or static option generating the view indices
+     * @param callable|Cache\GroupBy|null               $groupBy          The option or static option generating the group names
+     * @param array|callable|Cache\ChoiceAttr|null      $attr             The option or static option generating the HTML attributes
      */
     public function createView(ChoiceListInterface $list, $preferredChoices = null, $label = null, $index = null, $groupBy = null, $attr = null)
     {