+++ src/Symfony/Component/Form/ChoiceList/Factory/PropertyAccessDecorator.php
@@ -59,13 +59,17 @@ class PropertyAccessDecorator implements ChoiceListFactoryInterface
     /**
      * {@inheritdoc}
      *
-     * @param callable|string|PropertyPath|null $value The callable or path for
-     *                                                 generating the choice values
+     * @param callable|string|PropertyPath|null $value  The callable or path for
+     *                                                  generating the choice values
+     * @param callable|string|PropertyPath|null $filter The callable or path for
+     *                                                  filtering the choices
      *
      * @return ChoiceListInterface The choice list
      */
-    public function createListFromChoices(iterable $choices, $value = null)
+    public function createListFromChoices(iterable $choices, $value = null/*, $filter = null*/)
     {
+        $filter = \func_num_args() > 2 ? func_get_arg(2) : null;
+
         if (\is_string($value)) {
             $value = new PropertyPath($value);
         }
@@ -81,19 +85,34 @@ class PropertyAccessDecorator implements ChoiceListFactoryInterface
             };
         }
 
-        return $this->decoratedFactory->createListFromChoices($choices, $value);
+        if (\is_string($filter)) {
+            $filter = new PropertyPath($filter);
+        }
+
+        if ($filter instanceof PropertyPath) {
+            $accessor = $this->propertyAccessor;
+            $filter = static function ($choice) use ($accessor, $filter) {
+                return (\is_object($choice) || \is_array($choice)) && $accessor->getValue($choice, $filter);
+            };
+        }
+
+        return $this->decoratedFactory->createListFromChoices($choices, $value, $filter);
     }
 
     /**
      * {@inheritdoc}
      *
-     * @param callable|string|PropertyPath|null $value The callable or path for
-     *                                                 generating the choice values
+     * @param callable|string|PropertyPath|null $value  The callable or path for
+     *                                                  generating the choice values
+     * @param callable|string|PropertyPath|null $filter The callable or path for
+     *                                                  filtering the choices
      *
      * @return ChoiceListInterface The choice list
      */
-    public function createListFromLoader(ChoiceLoaderInterface $loader, $value = null)
+    public function createListFromLoader(ChoiceLoaderInterface $loader, $value = null/*, $filter = null*/)
     {
+        $filter = \func_num_args() > 2 ? func_get_arg(2) : null;
+
         if (\is_string($value)) {
             $value = new PropertyPath($value);
         }
@@ -109,7 +128,18 @@ class PropertyAccessDecorator implements ChoiceListFactoryInterface
             };
         }
 
-        return $this->decoratedFactory->createListFromLoader($loader, $value);
+        if (\is_string($filter)) {
+            $filter = new PropertyPath($filter);
+        }
+
+        if ($filter instanceof PropertyPath) {
+            $accessor = $this->propertyAccessor;
+            $filter = static function ($choice) use ($accessor, $filter) {
+                return (\is_object($choice) || \is_array($choice)) && $accessor->getValue($choice, $filter);
+            };
+        }
+
+        return $this->decoratedFactory->createListFromLoader($loader, $value, $filter);
     }
 
     /**