+++ src/Symfony/Component/Form/Extension/Core/Type/ChoiceType.php
@@ -15,6 +15,7 @@ use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\ChoiceList\ChoiceListInterface;
 use Symfony\Component\Form\ChoiceList\Factory\Cache\ChoiceAttr;
 use Symfony\Component\Form\ChoiceList\Factory\Cache\ChoiceFieldName;
+use Symfony\Component\Form\ChoiceList\Factory\Cache\ChoiceFilter;
 use Symfony\Component\Form\ChoiceList\Factory\Cache\ChoiceLabel;
 use Symfony\Component\Form\ChoiceList\Factory\Cache\ChoiceLoader;
 use Symfony\Component\Form\ChoiceList\Factory\Cache\ChoiceValue;
@@ -40,6 +41,7 @@ use Symfony\Component\Form\FormInterface;
 use Symfony\Component\Form\FormView;
 use Symfony\Component\OptionsResolver\Options;
 use Symfony\Component\OptionsResolver\OptionsResolver;
+use Symfony\Component\PropertyAccess\PropertyPath;
 
 class ChoiceType extends AbstractType
 {
@@ -52,6 +54,17 @@ class ChoiceType extends AbstractType
                 new DefaultChoiceListFactory()
             )
         );
+
+        // BC, to be removed in 6.0
+        if ($this->choiceListFactory instanceof CachingFactoryDecorator) {
+            return;
+        }
+
+        $ref = new \ReflectionMethod($this->choiceListFactory, 'createListFromChoices');
+
+        if ($ref->getNumberOfParameters() < 3) {
+            trigger_deprecation('symfony/form', '5.1', 'Not defining a third parameter "callable|null $filter" in "%s::%s()" is deprecated.', $ref->class, $ref->name);
+        }
     }
 
     /**
@@ -307,6 +320,7 @@ class ChoiceType extends AbstractType
             'multiple' => false,
             'expanded' => false,
             'choices' => [],
+            'choice_filter' => null,
             'choice_loader' => null,
             'choice_label' => null,
             'choice_name' => null,
@@ -332,6 +346,7 @@ class ChoiceType extends AbstractType
         $resolver->setAllowedTypes('choices', ['null', 'array', '\Traversable']);
         $resolver->setAllowedTypes('choice_translation_domain', ['null', 'bool', 'string']);
         $resolver->setAllowedTypes('choice_loader', ['null', 'Symfony\Component\Form\ChoiceList\Loader\ChoiceLoaderInterface', ChoiceLoader::class]);
+        $resolver->setAllowedTypes('choice_filter', ['null', 'callable', 'string', PropertyPath::class, ChoiceFilter::class]);
         $resolver->setAllowedTypes('choice_label', ['null', 'bool', 'callable', 'string', 'Symfony\Component\PropertyAccess\PropertyPath', ChoiceLabel::class]);
         $resolver->setAllowedTypes('choice_name', ['null', 'callable', 'string', 'Symfony\Component\PropertyAccess\PropertyPath', ChoiceFieldName::class]);
         $resolver->setAllowedTypes('choice_value', ['null', 'callable', 'string', 'Symfony\Component\PropertyAccess\PropertyPath', ChoiceValue::class]);
@@ -396,14 +411,19 @@ class ChoiceType extends AbstractType
         if (null !== $options['choice_loader']) {
             return $this->choiceListFactory->createListFromLoader(
                 $options['choice_loader'],
-                $options['choice_value']
+                $options['choice_value'],
+                $options['choice_filter']
             );
         }
 
         // Harden against NULL values (like in EntityType and ModelType)
         $choices = null !== $options['choices'] ? $options['choices'] : [];
 
-        return $this->choiceListFactory->createListFromChoices($choices, $options['choice_value']);
+        return $this->choiceListFactory->createListFromChoices(
+            $choices,
+            $options['choice_value'],
+            $options['choice_filter']
+        );
     }
 
     private function createChoiceListView(ChoiceListInterface $choiceList, array $options)